[fix]: Fix azure connection and clean up test (#150)
issue: #148

+++ cpp/include/milvus-storage/packed/column_group_writer.h
@@ -52,7 +52,7 @@ class ColumnGroupWriter {
   ColumnGroup column_group_;
   int flushed_batches_;
   int flushed_count_;
-  int64_t flushed_rows_;
+  int flushed_rows_;
 };
 
 }  // namespace milvus_storage
